# docs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skafolio API",
    "description": "Digital Business Card Platform API - Fast, secure, and feature-rich endpoints for card management, orders, payments, and analytics.",
    "version": "1.0.0",
    "contact": {
      "name": "Skafolio Support",
      "email": "support@skafolio.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "https://skafolio.vercel.app/api",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/payfast/ipn": {
      "post": {
        "tags": ["Payments"],
        "summary": "PayFast IPN webhook",
        "description": "Receives payment notifications from PayFast. Public endpoint with signature validation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPN processed successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
...
  }
}